LayoutTests/js/Number-isNaN-expected.txt

Test the basic behaviors of Number.isNaN()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Number.hasOwnProperty("isNaN") is true
PASS typeof Number.isNaN is "function"
PASS Number.isNaN !== isNaN is true
PASS Number.isNaN.length is 1
PASS Number.isNaN.name is "isNaN"
PASS Object.getOwnPropertyDescriptor(Number, "isNaN").configurable is true
PASS Object.getOwnPropertyDescriptor(Number, "isNaN").enumerable is false
PASS Object.getOwnPropertyDescriptor(Number, "isNaN").writable is true
PASS Number.isNaN() is false
PASS Number.isNaN(NaN) is true
PASS Number.isNaN(undefined) is false
PASS Number.isNaN(0) is false
PASS Number.isNaN(-0) is false
PASS Number.isNaN(1) is false
PASS Number.isNaN(-1) is false
PASS Number.isNaN(42) is false
PASS Number.isNaN(123.5) is false
PASS Number.isNaN(-123.5) is false
PASS Number.isNaN(Number.MAX_VALUE) is false
PASS Number.isNaN(Number.MIN_VALUE) is false
PASS Number.isNaN(Number.MAX_SAFE_INTEGER) is false
PASS Number.isNaN(Number.MIN_SAFE_INTEGER) is false
PASS Number.isNaN(Math.PI) is false
PASS Number.isNaN(Math.E) is false
PASS Number.isNaN(Infinity) is false
PASS Number.isNaN(-Infinity) is false
PASS Number.isNaN(null) is false
PASS Number.isNaN({}) is false
PASS Number.isNaN({ webkit: "awesome" }) is false
PASS Number.isNaN([]) is false
PASS Number.isNaN([123]) is false
PASS Number.isNaN([1,1]) is false
PASS Number.isNaN([NaN]) is false
PASS Number.isNaN("") is false
PASS Number.isNaN("1") is false
PASS Number.isNaN("x") is false
PASS Number.isNaN("NaN") is false
PASS Number.isNaN("Infinity") is false
PASS Number.isNaN(true) is false
PASS Number.isNaN(false) is false
PASS Number.isNaN(function(){}) is false
PASS Number.isNaN(isNaN) is false
PASS Number.isNaN(Symbol()) is false
PASS Number.isNaN(objectWithNumberValueOf) is false
PASS Number.isNaN(objectWithNaNValueOf) is false
PASS Number.isNaN(objectRecordConversionCalls) is false
PASS objectRecordConversionCalls.callList.length is 0
PASS successfullyParsed is true

TEST COMPLETE
